backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
motor
networkx
python-dotenv
joblib
scikit-learn
numpy
pydantic-settings
pandas
httpx
tqdm
lightgbm
catboost